+++ src/views/CommonView.tsx
@@ -5,6 +5,8 @@ import { AppFeature } from '~core/auth/types';
 import { Row } from '~components/Layout';
 import { OriginalLogo } from '~components/KonturLogo/KonturLogo';
 import { userResourceAtom } from '~core/auth';
+import { currentAppPropertiesResourceAtom } from '~core/shared_state/currentApplication';
+import { useBrowserTabUpdateEffect } from '~utils/hooks/useBrowserTabUpdate';
 import type { AvailableRoutesAtom, CurrentRouteAtom } from '~core/router';
 import type { PropsWithChildren } from 'react';
 
@@ -24,6 +26,9 @@ export function CommonView({
 }>) {
   const [{ data, loading }] = useAtom(userResourceAtom);
   const userModel = data && !loading ? data : null;
+  const [{ data: appParams }] = useAtom(currentAppPropertiesResourceAtom);
+
+  useBrowserTabUpdateEffect(appParams?.faviconUrl);
 
   useEffect(() => {
     if (userModel?.hasFeature(AppFeature.INTERCOM)) {
